+++ Library/include/CSP/Systems/Spaces/SpaceSystem.h
@@ -56,6 +56,7 @@ namespace csp::systems
 {
 
 class UserSystem;
+class MultiplayerSystem;
 
 /// @ingroup Space System
 /// @brief Public facing system that allows interfacing with Magnopus Connected Services' concept of a Group.
@@ -380,6 +381,9 @@ public:
         const csp::common::Optional<csp::common::Array<csp::common::String>>& MemberGroupIds, bool ShallowCopy, NullResultCallback Callback);
 
     ///@}
+    // This is required due to a circular dependency between SpaceSystem and MultiplayerSystem.
+    // This will be broken when we move enter space logic into RealtimeEngine.
+    CSP_NO_EXPORT void SetMultiplayerSystem(csp::systems::MultiplayerSystem& MultiplayerSystem);
 
     /// @brief The callback for receiving an alert when an async operation is completed.
     /// Currently this callback is only being used for the DuplicateSpaceAsync operation.
@@ -439,6 +443,9 @@ private:
         const std::shared_ptr<SpaceResult>& Space, const csp::systems::BufferAssetDataSource& Data);
     std::function<async::task<NullResult>()> BulkInviteUsersToSpaceIfNeccesary(
         SpaceSystem* SpaceSystem, const std::shared_ptr<SpaceResult>& Space, const csp::common::Optional<InviteUserRoleInfoCollection>& InviteUsers);
+    // This currently checks if the default scope has leader election enabled, if so will enable server-side leader election in the
+    // OnlineRealtimeEngine and register the scope to keep track of its leader.
+    std::function<async::task<SpaceResult>(const SpaceResult& SpaceResult)> RegisterScopesInSpace(csp::common::IRealtimeEngine* RealtimeEngine);
 
     // EnterSpace Continuations
     auto AddUserToSpaceIfNecessary(SpaceResultCallback Callback, SpaceSystem& SpaceSystem);
@@ -449,6 +456,8 @@ private:
     csp::services::ApiBase* GroupAPI;
     csp::services::ApiBase* SpaceAPI;
     Space CurrentSpace;
+
+    csp::systems::MultiplayerSystem* MultiplayerSystem;
 };
 
 } // namespace csp::systems