+++ MultiplayerTestRunner/src/RunnableTests/LeaderElection.h
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2025 Magnopus LLC
+
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#pragma once
+
+namespace csp::multiplayer
+{
+class OnlineRealtimeEngine;
+}
+
+namespace LeaderElection
+{
+/*
+ * Test that waits to be made the leader of a scope, and then disconnects, re-triggering leadership election.
+ */
+void RunTest(csp::multiplayer::OnlineRealtimeEngine& RealtimeEngine);
+} // namespace LeaderElection